Changed include/src paths to be behind an abdeveng/core/* include path.
