+++ core/core/core.vcxproj
@@ -10,6 +10,16 @@
       <Platform>x64</Platform>
     </ProjectConfiguration>
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="include\abdeveng\core\test\common\case.h" />
+    <ClInclude Include="include\abdeveng\core\test\common\evaluator.h" />
+    <ClInclude Include="include\abdeveng\core\test\common\harness.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="src\abdeveng\core\test\common\case.cpp" />
+    <ClCompile Include="src\abdeveng\core\test\common\evaluator.cpp" />
+    <ClCompile Include="src\abdeveng\core\test\common\harness.cpp" />
+  </ItemGroup>
   <PropertyGroup Label="Globals">
     <ProjectGuid>{D8683758-AAC9-4C64-A180-75FB7212010E}</ProjectGuid>
     <RootNamespace>core</RootNamespace>
@@ -38,16 +48,17 @@
   </ImportGroup>
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
-    <IncludePath>$(VC_IncludePath);$(WindowsSDK_IncludePath);$(ProjectDir)\include</IncludePath>
+    <IncludePath>$(VC_IncludePath);$(WindowsSDK_IncludePath);$(ProjectDir)/include</IncludePath>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <IncludePath>$(VC_IncludePath);$(WindowsSDK_IncludePath);$(ProjectDir)\include</IncludePath>
+    <IncludePath>$(VC_IncludePath);$(WindowsSDK_IncludePath);$(ProjectDir)/include</IncludePath>
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <ClCompile>
       <WarningLevel>Level3</WarningLevel>
       <Optimization>Disabled</Optimization>
       <SDLCheck>true</SDLCheck>
+      <ObjectFileName>$(IntDir)/%(RelativeDir)/</ObjectFileName>
     </ClCompile>
     <Link>
       <GenerateDebugInformation>true</GenerateDebugInformation>
@@ -60,6 +71,7 @@
       <FunctionLevelLinking>true</FunctionLevelLinking>
       <IntrinsicFunctions>true</IntrinsicFunctions>
       <SDLCheck>true</SDLCheck>
+      <ObjectFileName>$(IntDir)/%(RelativeDir)/</ObjectFileName>
     </ClCompile>
     <Link>
       <GenerateDebugInformation>true</GenerateDebugInformation>
@@ -67,17 +79,6 @@
       <OptimizeReferences>true</OptimizeReferences>
     </Link>
   </ItemDefinitionGroup>
-  <ItemGroup>
-    <ClInclude Include="include\service\logger.h" />
-    <ClInclude Include="include\test\common\evaluator.h" />
-    <ClInclude Include="include\test\common\case.h" />
-    <ClInclude Include="include\test\common\harness.h" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="src\test\common\evaluator.cpp" />
-    <ClCompile Include="src\test\common\case.cpp" />
-    <ClCompile Include="src\test\common\harness.cpp" />
-  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
